Standalone launch file for rviz
